requirements.txt
absl-py==0.4.0
astor==0.7.1
boto==2.49.0
boto3==1.7.66
botocore==1.10.66
bz2file==0.98
certifi==2018.4.16
chardet==3.0.4
Cython==0.28.4
docutils==0.14
gast==0.2.0
gensim==3.5.0
grpcio==1.14.1
h5py==2.8.0
idna==2.7
jmespath==0.9.3
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
Markdown==2.6.11
numpy==1.14.5
protobuf==3.6.1
python-dateutil==2.7.3
PyYAML==3.13
requests==2.19.1
s3transfer==0.1.13
scipy==1.1.0
six==1.11.0
smart-open==1.6.0
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
tweet-preprocessor==0.5.0
urllib3==1.23
Werkzeug==0.14.1
